DEF[#56813]: 在设置了任务结束时间后 任务还在执行
缺陷【#56114】： 用户越权告警
缺陷[#56118]: 【视频控制】视频球机概率性出现无法控制的情况，提示“相机不可控！”
DEF[#56101]: 巡视系统下发周期任务，个别周期任务要执行的所有任务数据不显示在任务日历中
DEF[#55990]: 一键顺控调用莱达算法传mp4文件
DEF[#55990]:联动任务下发失败
缺陷【#55870】： 摄像机任务，下载的报告中图片不是结果图
DEF[#55734]: 机器人周期和间隔任务不能结束
DEF[#55734]: 【500检测】摄像机台账，录像机字段值显示id，应显示名称
DEF[#55909]: 机器人周期任务下发重复
缺陷【#55870】： 摄像机任务，下载的报告中图片不是结果图
DEF[#55909]: 【500检测】下发大量（大于70个）机器人与视频的混合测点任务，当有告警弹框出现时，提示“更新数据失败”
DEF[#55911]: 【上级系统接口】巡视系统向上级系统上报巡视结果type61报文，Code字段为空，应为变电站编码值
DEF[#55838]: 【500检测】下发周期任务，缺少间隔数量、间隔结束时间

See merge request iotcenter/IotCenter!21

(cherry picked from commit fe4fc0b444d9b0563517d6a2d3bd81485f9f2d10)

0f6c54c1 缺陷【#55868】：视频监控页面业务员调用了预置位删除新增,提示越权

+++ iot-center-platform/src/main/java/com/yjh/platform/module/device/controller/TStdDevicemeteController.java
@@ -13,6 +13,7 @@ import java.util.HashMap;
 import java.util.List;
 
 import io.swagger.annotations.*;
+import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -80,7 +81,13 @@ public class TStdDevicemeteController {
     public Result delete(@RequestParam(value = "deviceMeteId") Long deviceMeteId) {
         Result result = new Result();
         try {
-            result.setData(tStdDevicemeteService.deleteByPrimaryId(deviceMeteId));
+            // 当前测点配置巡视点 ？ 不允许删除 ： 删除逻辑
+            List<Long> haveList = tStdDevicemeteDao.selectHave(deviceMeteId);
+            if (CollectionUtils.isNotEmpty(haveList)){
+                result.setMessage(209,"测点已配置为巡视点,操作无法生效");
+            }else {
+                result.setData(tStdDevicemeteService.deleteByPrimaryId(deviceMeteId));
+            }
         } catch (BusinessException e) {
             result.setMessage(ResultCodeEnum.SYSTEMERROR.getCode(), e.getMessage());
             log.error("标准设备测点删除异常:", e);
@@ -176,7 +183,7 @@ public class TStdDevicemeteController {
 
     @ApiOperation(value = "分页查询")
     @RequestMapping(value = "/selectByPage", method = RequestMethod.POST)
-    @Logs(title = "查询标准设备测点",content = "根据用户传递的参数查询标准设备测点",logType = 1,authority = "1234")
+    @Logs(title = "查询标准设备测点",content = "根据用户传递的参数查询标准设备测点",logType = 1,authority = "1234,1235")
     public Result selectByPage(@RequestBody TStdDeviceMeteDetail tStdDeviceMeteDetail
                               ) {
         Result result = new Result();
